+++ framework/src/main/java/org/checkerframework/framework/flow/CFAbstractTransfer.java
@@ -205,24 +205,6 @@ public abstract class CFAbstractTransfer<
     return analysis.createAbstractValue(at);
   }
 
-  /**
-   * Returns an abstract value with the given {@code type} and the annotations from {@code
-   * annotatedValue}.
-   *
-   * @param type the type to return
-   * @param annotatedValue the annotations to return
-   * @return an abstract value with the given {@code type} and the annotations from {@code
-   *     annotatedValue}
-   * @deprecated use {@link #getWidenedValue} or {@link #getNarrowedValue}
-   */
-  @Deprecated // 2020-10-02
-  protected V getValueWithSameAnnotations(TypeMirror type, V annotatedValue) {
-    if (annotatedValue == null) {
-      return null;
-    }
-    return analysis.createAbstractValue(annotatedValue.getAnnotations(), type);
-  }
-
   /** The fixed initial store. */
   private S fixedInitialStore = null;